Extracts emails from messy ASCII text
Simple app that will accept file or paste ASCII input, and use a RegEx
matching function to extract email addressed, so they can be displayed
in a convenient list. This is sometimes helpful to preprocess email
lists for upload to create contacts for bulk email services.

This should only be done to extra email addresses for an established
base of contacts for which have given permission for you to send them
emails.